# orchest-ui/swagger.json
{
"openapi": "3.1.0",
"info": {
"title": "OrchesT",
"description": "REST interface for interacting with Orchest Engine",
"contact": {
"name": "DOT Team Tyrell",
"email": "teamtyrell@telekom-digital.com"
},
"license": {
"name": "Apache License, Version 2.0"
},
"version": "1.0.0",
"x-api-category": "other"
},
"servers": [
{
"url": "http://localhost:6200/orchest/"
},
{
"url": "https://api-orchest.oneapp.dev.hal.telekom.de/orchest/"
},
{
"url": "https://api-orchest.oneapp.teststable.hal.telekom.de/orchest/"
}
],
...
"paths": {
...
"/processStates": {
"get": {
"tags": [
"process-state-api"
],
"operationId": "getProcessStates",
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/ProcessState"
}
}
}
}
}
}
},
"put": {
"tags": [
"process-state-api"
],
"operationId": "upsertProcessState",
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
},
"required": true
},
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
}
}
}
},
"post": {
"tags": [
"process-state-api"
],
"operationId": "addProcessState",
"requestBody": {
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
},
"required": true
},
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
}
}
}
}
},
...
"/userTasks/process/{processInstanceId}": {
"get": {
"tags": [
"User Tasks"
],
"summary": "List tasks for a process instance",
"operationId": "getTasksForProcessInstance",
"parameters": [
{
"name": "processInstanceId",
"in": "path",
"description": "Process instance ID",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "page",
"in": "query",
"description": "Zero-based page index",
"required": false,
"schema": {
"type": "integer",
"format": "int32",
"default": 0
}
},
{
"name": "size",
"in": "query",
"description": "Page size",
"required": false,
"schema": {
"type": "integer",
"format": "int32",
"default": 10
}
},
{
"name": "sort",
"in": "query",
"description": "Sort field with direction prefix (+ASC / -DESC)",
"required": false,
"schema": {
"type": "string",
"default": "-createdAt"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/PageUserTaskDTO"
}
}
}
}
}
}
},
...
"/processStates/{processDefinitionId}": {
"get": {
"tags": [
"process-state-api"
],
"operationId": "getProcessStateForProcessId",
"parameters": [
{
"name": "processDefinitionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
}
}
}
},
"delete": {
"tags": [
"process-state-api"
],
"operationId": "removeProcessState",
"parameters": [
{
"name": "processDefinitionId",
"in": "path",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "OK",
"content": {
"*/*": {
"schema": {
"$ref": "#/components/schemas/ProcessState"
}
}
}
}
}
}
},
...
},
"components": {
"schemas": {
...
"ProcessState": {
"type": "object",
"properties": {
"id": {
"type": "string"
},
"processId": {
"type": "string"
},
"status": {
"type": "boolean"
}
}
},
...
"UserTaskDTO": {
"type": "object",
"properties": {
"taskId": {
"type": "string"
},
"processInstanceId": {
"type": "string"
},
"processDefinitionId": {
"type": "string"
},
"activityId": {
"type": "string"
},
"taskName": {
"type": "string"
},
"state": {
"type": "string"
},
"assignee": {
"type": "string"
},
"candidateUsers": {
"type": "array",
"items": {
"type": "string"
}
},
"candidateGroups": {
"type": "array",
"items": {
"type": "string"
}
},
"claimedBy": {
"type": "string"
},
"dueDate": {
"type": "string"
},
"followUpDate": {
"type": "string"
},
"formKey": {
"type": "string"
},
"variables": {
"type": "object",
"additionalProperties": {}
},
"createdAt": {
"type": "string"
},
"claimedAt": {
"type": "string"
},
"completedAt": {
"type": "string"
}
}
},
...
"PageUserTaskDTO": {
"type": "object",
"properties": {
"totalPages": {
"type": "integer",
"format": "int32"
},
"totalElements": {
"type": "integer",
"format": "int64"
},
"pageable": {
"$ref": "#/components/schemas/PageableObject"
},
"first": {
"type": "boolean"
},
"last": {
"type": "boolean"
},
"numberOfElements": {
"type": "integer",
"format": "int32"
},
"size": {
"type": "integer",
"format": "int32"
},
"content": {
"type": "array",
"items": {
"$ref": "#/components/schemas/UserTaskDTO"
}
},
"number": {
"type": "integer",
"format": "int32"
},
"sort": {
"$ref": "#/components/schemas/SortObject"
},
"empty": {
"type": "boolean"
}
}
},
"PageableObject": {
"type": "object",
"properties": {
"paged": {
"type": "boolean"
},
"pageNumber": {
"type": "integer",
"format": "int32"
},
"pageSize": {
"type": "integer",
"format": "int32"
},
"unpaged": {
"type": "boolean"
},
"offset": {
"type": "integer",
"format": "int64"
},
"sort": {
"$ref": "#/components/schemas/SortObject"
}
}
},
"SortObject": {
"type": "object",
"properties": {
"sorted": {
"type": "boolean"
},
"unsorted": {
"type": "boolean"
},
"empty": {
"type": "boolean"
}
}
},
...
}
}
}